outsources usdb id file parser with logging and validation to separate module as pyside imports seem to harm tests in ci/cd
